chore: remove unused pkg, interop vitest
