+++ packages/frontend/package.json
@@ -18,7 +18,6 @@
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/themes": "^2.0.0",
     "@svgr/webpack": "^8.1.0",
-    "@tremor/react": "^3.11.1",
     "@types/ramda": "^0.29.7",
     "clsx": "^2.0.0",
     "emoji-picker-react": "^4.4.7",
