+++ test/setup-file.ts
@@ -1,6 +1,31 @@
 import { expect } from 'vitest';
+import { ZodError } from 'zod';
 
-import { toBeSafe } from '@test/schema';
+export function toBeSafe<T extends any>(
+  received: { success: true; data: T; error?: ZodError },
+  expected?: T,
+) {
+  // @ts-ignore
+  const { isNot } = this;
+  return {
+    // 请勿根据 isNot 参数更改你的 "pass" 值，Vitest 为你做了这件事情
+    pass:
+      expected !== undefined
+        ? received.success && received.data === expected
+        : received.success,
+    message: () => `${received.data} is${isNot ? ' not' : ''} safe`,
+  };
+}
+
+interface CustomMatchers<R = unknown> {
+  toBeSafe(): R;
+}
+
+declare module 'vitest' {
+  interface Assertion<T = any> extends CustomMatchers<T> {}
+
+  interface AsymmetricMatchersContaining extends CustomMatchers {}
+}
 
 expect.extend({ toBeSafe });
